[Refactor] term, application.yml 변경 (#1)
- term : nullable 옵션, length 지정, primitive type/정수형으로 변경
- application.yml : ddl-auto 옵션 validate로 변경
